rag-series/shared/requirements.txt
# Shared dependencies for RAG series LLM providers
# Blog 2.3: From Retrieval to Answers - The Full RAG Pipeline

# LLM API clients
openai>=1.0.0
anthropic>=0.3.0

# Configuration and environment management
python-dotenv>=0.19.0

# Async support
asyncio-throttle>=1.0.0

# Optional dependencies for token counting
tiktoken>=0.5.0  # For accurate OpenAI token counting

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.20.0